requirements.txt
awesome-slugify==1.6.5
Babel==1.3
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.1
bs4==0.0.1
certifi==2015.11.20.1
cffi==1.7.0
coverage==4.0.3
cryptography==1.4
DAWG-Python==0.7.2
decorator==4.0.10
diff-match-patch==20121119
Django==1.10.3
django-admin-bootstrapped==2.5.7
django-angular==0.8.4
django-appconf==1.0.1
django-autocomplete-light==3.1.6
django-bootstrap-pagination==1.6.1
django-bootstrap3==7.0.1
django-braces==1.8.1
django-ckeditor==5.0.3
django-compressor==1.6
django-debug-panel==0.8.3
django-debug-toolbar==1.4
django-extensions==1.7.5
django-extra-views==0.6.4
django-filer==1.2.2
django-flat-theme==1.1.3
-e git+https://github.com/django-haystack/django-haystack.git@99e5a10ae8535a7883ff95d2e83fa4f127db546d#egg=django_haystack
django-htmlmin==0.9.1
django-import-export==0.4.5
django-mptt==0.8.6
django-oscar==1.3
django-oscar-accounts==0.3
django-phonenumber-field==1.1.0
django-polymorphic==1.0.2
django-selenium==0.9.8
django-smart-load-tag==0.3.2
django-smartagent==0.1.1
django-suit==0.2.15
django-tables2==1.0.6
django-treebeard==4.1.0
django-widget-tweaks==1.3
docopt==0.6.2
easy-thumbnails==2.3
EasyProcess==0.2.2
ecdsa==0.13
enum34==1.1.6
Fabric==1.10.2
factory-boy==2.4.1
feedparser==5.2.1
FeinCMS==1.11.4
google-api-python-client==1.5.1
gunicorn==19.6.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.1
inflect==0.2.5
ipaddress==1.0.16
ipython==5.1.0
ipython-genutils==0.1.0
memcache-toolbar==0.5.6
mock==1.0.1
oauth2client==3.0.0
paramiko==1.16.0
path.py==8.1.2
pathlib2==2.1.0
pexpect==4.2.1
phonenumbers==6.3.0
phonenumberslite==7.5.1
pickleshare==0.7.4
Pillow==2.7.0
prompt-toolkit==1.0.7
psycopg2==2.6.1
ptyprocess==0.5.1
purl==1.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyOpenSSL==16.0.0
pysolr==3.3.3
pytest-spec==0.2.24
python-dateutil==2.1
python-memcached==1.57
python-wd-parallel==0.1.1
pytz==2015.7
PyVirtualDisplay==0.2
PyYAML==3.11
regex==2016.1.10
requests==2.9.1
rsa==3.4.2
selenium==2.53.5
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sorl-thumbnail==12.4a1
sqlparse==0.1.18
tablib==0.11.2
texttable==0.8.4
tornado==4.3
traitlets==4.2.2
Unidecode==0.4.18
uritemplate==0.6
wcwidth==0.1.7
webencodings==0.5
Whoosh==2.7.4
